Rename various utility modules to _util.ts
