+++ bokehjs/test/unit/models/glyphs/_util.ts
