+++ bokehjs/test/unit/models/glyphs/image_rgba.ts
@@ -1,6 +1,6 @@
 import {expect} from "assertions"
 
-import {create_glyph_view} from "./glyph_utils"
+import {create_glyph_view} from "./_util"
 import {ImageRGBA} from "@bokehjs/models/glyphs/image_rgba"
 import {ndarray} from "@bokehjs/core/util/ndarray"
 import {NumberArray} from '@bokehjs/core/types'
