+++ bokehjs/test/unit/models/tools/edit/_util.ts
