+++ bokehjs/test/unit/models/tools/edit/box_edit_tool.ts
@@ -12,7 +12,7 @@ import {GlyphRenderer} from "@bokehjs/models/renderers/glyph_renderer"
 import {ColumnDataSource} from "@bokehjs/models/sources/column_data_source"
 import {BoxEditTool, BoxEditToolView} from "@bokehjs/models/tools/edit/box_edit_tool"
 
-import {make_pan_event, make_tap_event, make_move_event, make_key_event} from "./utils"
+import {make_pan_event, make_tap_event, make_move_event, make_key_event} from "./_util"
 
 export interface BoxEditTestCase {
   data: {[key: string]: (number | null)[]}
